chore: update Rollup configuration and commitlint rules
- Added a new Rollup configuration file with alias support for better module resolution.
- Updated the build script in package.json to use the new Rollup configuration file.
- Modified commitlint configuration to disable body-max-line-length rule.
- Added @rollup/plugin-alias as a dev dependency for module aliasing.
