--- rollup.config.js
@@ -1,15 +1,0 @@
-import commonjs from "@rollup/plugin-commonjs";
-import { nodeResolve } from "@rollup/plugin-node-resolve";
-
-const config = {
-  input: "src/index.js",
-  output: {
-    esModule: true,
-    file: "dist/index.js",
-    format: "es",
-    sourcemap: true,
-  },
-  plugins: [commonjs(), nodeResolve({ preferBuiltins: true })],
-};
-
-export default config;
